+++ rust/hyperlane-core/src/accumulator/mod.rs
@@ -3,8 +3,8 @@ use sha3::{Digest, Keccak256};
 
 use crate::H256;
 
-/// Utilities for manipulating proofs to reflect historic merkle trees.
-pub mod historic;
+/// Utilities for manipulating proofs to reflect sparse merkle trees.
+pub mod sparse;
 /// A lightweight incremental merkle, suitable for running on-chain. Stores O
 /// (1) data
 pub mod incremental;
